Iteration 3.
