+++ src/main/java/com/acme/edu/BufferState.java
@@ -1,0 +1,8 @@
+package com.acme.edu;
+
+public enum BufferState {
+    BYTE,
+    INT,
+    STR,
+    NONE
+}
